kt/src/test/kotlin/io/github/ephemient/aoc2020/Day7Test.kt
package io.github.ephemient.aoc2020

import kotlin.test.Test
import kotlin.test.assertEquals

class Day7Test {
    @Test
    fun `part 1 examples`() {
        assertEquals(4, Day7(example1).part1())
    }

    @Test
    fun `part 2 examples`() {
        assertEquals(32, Day7(example1).part2())
        assertEquals(126, Day7(example2).part2())
    }

    companion object {
        private val example1 = listOf(
            "light red bags contain 1 bright white bag, 2 muted yellow bags.",
            "dark orange bags contain 3 bright white bags, 4 muted yellow bags.",
            "bright white bags contain 1 shiny gold bag.",
            "muted yellow bags contain 2 shiny gold bags, 9 faded blue bags.",
            "shiny gold bags contain 1 dark olive bag, 2 vibrant plum bags.",
            "dark olive bags contain 3 faded blue bags, 4 dotted black bags.",
            "vibrant plum bags contain 5 faded blue bags, 6 dotted black bags.",
            "faded blue bags contain no other bags.",
            "dotted black bags contain no other bags.",
        )

        private val example2 = listOf(
            "shiny gold bags contain 2 dark red bags.",
            "dark red bags contain 2 dark orange bags.",
            "dark orange bags contain 2 dark yellow bags.",
            "dark yellow bags contain 2 dark green bags.",
            "dark green bags contain 2 dark blue bags.",
            "dark blue bags contain 2 dark violet bags.",
            "dark violet bags contain no other bags.",
        )
    }
}
